Add copyFile function
